Add Language Code Parameter
I have added the parameter LangCode for the function "Invoke-DHCPCheckup". This parameter is used in the "GetStrongUsers" function to change the list of $strongGroups so that the name of these groups matches a specific language used in the Active Directory.
I've added German to the list. English was already in the original.